fix: move state creation into prelink and put it on scope
In AngularJs directives are created using a factory method, hence the state at the directive method is kept for all instances of the same directive. This introduced new issue
where multiple instances of the directive linked to different elements would see the same state for init. I have fixed that by moving the init state to the scope, which is recreated each time directive is bind to an element, thus guaranteeing recreation of the state for each new linking.

Closes #298