fix: add property for clearAriaLabel in phoneInput
PhoneInput internally uses Input component which expects clearAriaLabel
as optional parameter while PhoneInput doesn't have any parameter for it
due to which we were not able to pass our localized label to it hence
resulting in accessibility issue. We added clearArialabel as a parameter
to PhoneInput and passed it to Input component.